Move the peripheral delegate out of the peripheral.
