+++ Sources/Peripheral/Peripheral.swift
@@ -8,15 +8,7 @@ import os.log
 /// A remote peripheral device.
 /// - This class acts as a wrapper around `CBPeripheral`.
 public class Peripheral {
-    
-    fileprivate class DelegateWrapper: NSObject {
-        fileprivate let context: PeripheralContext
         
-        init(context: PeripheralContext) {
-            self.context = context
-        }
-    }
-    
     private static let logger = Logger(
         subsystem: Bundle(for: Peripheral.self).bundleIdentifier ?? "",
         category: "peripheral"
@@ -60,7 +52,7 @@ public class Peripheral {
     private let context: PeripheralContext
     /// The delegate object that will receive `cbPeripheral`.
     /// - Note: We need to hold on to it because `cbPeripheral` has a weak reference to it.
-    private let cbPeripheralDelegate: DelegateWrapper
+    private let cbPeripheralDelegate: PeripheralDelegate
     
     public init(_ cbPeripheral: CBPeripheral) {
         self.cbPeripheral = cbPeripheral
@@ -68,7 +60,7 @@ public class Peripheral {
         // By reusing the cbPeripheralDelegate and context, we guarantee that we will enqueue calls to the peripheral
         // using the same context, and that won't lose any callbacks from the CBPeripheralDelegate.
         // This is important because we can create multiple Peripherals for a single cbPeripheral.
-        if let cbPeripheralDelegate = cbPeripheral.delegate as? DelegateWrapper {
+        if let cbPeripheralDelegate = cbPeripheral.delegate as? PeripheralDelegate {
             self.context = cbPeripheralDelegate.context
             self.cbPeripheralDelegate = cbPeripheralDelegate
             return
@@ -79,7 +71,7 @@ public class Peripheral {
         }
         
         self.context = PeripheralContext()
-        self.cbPeripheralDelegate = DelegateWrapper(context: self.context)
+        self.cbPeripheralDelegate = PeripheralDelegate(context: self.context)
         self.cbPeripheral.delegate = self.cbPeripheralDelegate
     }
     
@@ -219,168 +211,6 @@ public class Peripheral {
     func writeValue(_ data: Data, for descriptor: CBDescriptor) async throws {
         try await self.context.writeDescriptorValueExecutor.enqueue(withKey: descriptor.uuid) { [weak self] in
             self?.cbPeripheral.writeValue(data, for: descriptor)
-        }
-    }
-}
-
-// MARK: CBPeripheralDelegate
-
-extension Peripheral.DelegateWrapper: CBPeripheralDelegate {
-    private static var logger: Logger = {
-        Peripheral.logger
-    }()
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didReadRSSI RSSI: NSNumber, error: Error?) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: RSSI, error: error)
-                try await self.context.readRSSIExecutor.setWorkCompletedWithResult(result)
-            } catch {
-                Self.logger.error("Received ReadRSSI response without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didDiscoverServices error: Error?) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.discoverServiceExecutor.setWorkCompletedWithResult(result)
-            } catch {
-                Self.logger.warning("Received DiscoverServices response without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didDiscoverIncludedServicesFor service: CBService, error: Error?) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.discoverIncludedServicesExecutor.setWorkCompletedForKey(
-                    service.uuid, result: result
-                )
-            } catch {
-                Self.logger.warning("Received DiscoverIncludedServices response without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didDiscoverCharacteristicsFor service: CBService, error: Error?) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.discoverCharacteristicsExecutor.setWorkCompletedForKey(
-                    service.uuid, result: result
-                )
-            } catch {
-                Self.logger.warning("Received DiscoverCharacteristics result without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didUpdateValueFor characteristic: CBCharacteristic, error: Error?) {
-        
-        if characteristic.isNotifying {
-           
-           // characteristic.value is Data() and it will get trampled if allowed to run async.
-           self.context.characteristicValueUpdatedSubject.send( Characteristic(characteristic) )
-
-        }
-           
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.readCharacteristicValueExecutor.setWorkCompletedForKey(
-                    characteristic.uuid, result: result
-                )
-            } catch {
-                guard !characteristic.isNotifying else { return }
-                Self.logger.warning("Received UpdateValue result for characteristic without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didWriteValueFor characteristic: CBCharacteristic, error: Error?) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.writeCharacteristicValueExecutor.setWorkCompletedForKey(
-                    characteristic.uuid, result: result
-                )
-            } catch {
-                Self.logger.warning("Received WriteValue result for characteristic without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(
-        _ cbPeripheral: CBPeripheral,
-        didUpdateNotificationStateFor characteristic: CBCharacteristic,
-        error: Error?
-    ) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.setNotifyValueExecutor.setWorkCompletedForKey(
-                    characteristic.uuid, result: result
-                )
-            } catch {
-                Self.logger.warning("Received UpdateNotificationState result without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(
-        _ cbPeripheral: CBPeripheral,
-        didDiscoverDescriptorsFor characteristic: CBCharacteristic,
-        error: Error?
-    ) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.discoverDescriptorsExecutor.setWorkCompletedForKey(
-                    characteristic.uuid, result: result
-                )
-            } catch {
-                Self.logger.warning("Received DiscoverDescriptors result without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didUpdateValueFor descriptor: CBDescriptor, error: Error?) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.readDescriptorValueExecutor.setWorkCompletedForKey(
-                    descriptor.uuid, result: result
-                )
-            } catch {
-                Self.logger.warning("Received UpdateValue result for descriptor without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didWriteValueFor descriptor: CBDescriptor, error: Error?) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: (), error: error)
-                try await self.context.writeDescriptorValueExecutor.setWorkCompletedForKey(
-                    descriptor.uuid, result: result
-                )
-            } catch {
-                Self.logger.warning("Received WriteValue result for descriptor without a continuation")
-            }
-        }
-    }
-    
-    func peripheral(_ cbPeripheral: CBPeripheral, didOpen channel: CBL2CAPChannel?, error: Error?) {
-        Task {
-            do {
-                let result = CallbackUtils.result(for: channel, error: error)
-                try await self.context.openL2CAPChannelExecutor.setWorkCompletedWithResult(result)
-            } catch {
-                Self.logger.warning("Received OpenChannel result without a continuation")
-            }
         }
     }
 }
